gemini-root-cause/requirements.txt
streamlit
google-adk
python-dotenv